Minor bug fixes. Now DFS-MT works robustly and efficiently. Still need to fix falling piece zig-zag.

+++ ai/HolyHeuristic.cpp
@@ -13,7 +13,7 @@ namespace tetris
 			// --- Count the number of holes in board ---
 			int nHoles = countHoles(board);
 
-			return -nHoles;
+			return static_cast<float>(-nHoles);
 		}
 
 		std::unique_ptr<HeuristicBase> HolyHeuristic::clone() const
